Refactors scene select into two states so opening a scene through cli behaves a bit better
